+++ helpers.ts
@@ -5,8 +5,52 @@ import { ReadStream } from "tty";
 import { Configurations } from "./constants";
 import { EventEmitter } from "events";
 import * as crypto from "crypto";
+import progress = require("progress-stream");
+import filesize = require("filesize");
+import * as util from "util";
 
 const Table = require("cli-table");
+
+export function trackDownloadProgress(destinationStream: NodeJS.WritableStream, url: string): NodeJS.ReadableStream {
+	// \r for carriage return doesn't work on windows in node for some reason so we have to use it's hex representation \x1B[0G
+	let lastMessageSize = 0;
+	const carriageReturn = "\x1B[0G";
+	let timeElapsed = 0;
+
+	const isInteractiveTerminal = isInteractive();
+	const progressStream = progress({ time: 1000 }, (progress: any) => {
+		timeElapsed = progress.runtime;
+
+		if (timeElapsed >= 1) {
+			if (isInteractiveTerminal) {
+				this.$logger.write("%s%s", carriageReturn, Array(lastMessageSize + 1).join(" "));
+
+				const message = util.format("%sDownload progress ... %s | %s | %s/s",
+					carriageReturn,
+					Math.floor(progress.percentage) + "%",
+					filesize(progress.transferred),
+					filesize(progress.speed));
+
+				this.$logger.write(message);
+				lastMessageSize = message.length;
+			}
+		}
+	});
+
+	progressStream.on("finish", () => {
+		if (timeElapsed >= 1) {
+			const msg = `Download of ${url} completed.`;
+			if (isInteractiveTerminal) {
+				this.$logger.out("%s%s%s%s", carriageReturn, Array(lastMessageSize + 1).join(" "), carriageReturn, msg);
+			} else {
+				this.$logger.out(msg);
+			}
+		}
+	});
+
+	progressStream.pipe(destinationStream);
+	return progressStream;
+}
 
 export async function executeActionByChunks<T>(initialData: T[] | IDictionary<T>, chunkSize: number, elementAction: (element: T, key?: string | number) => Promise<any>): Promise<void> {
 	let arrayToChunk: (T | string)[];
