+++ http-client.ts
@@ -3,8 +3,6 @@ import { EOL } from "os";
 import * as helpers from "./helpers";
 import * as zlib from "zlib";
 import * as util from "util";
-import progress = require("progress-stream");
-import filesize = require("filesize");
 import { HttpStatusCodes } from "./constants";
 import * as request from "request";
 
@@ -133,8 +131,6 @@ export class HttpClient implements Server.IHttpClient {
 							this.setResponseResult(promiseActions, timerId, { response });
 						});
 
-						pipeTo = this.trackDownloadProgress(pipeTo);
-
 						responseStream.pipe(pipeTo);
 					} else {
 						const data: string[] = [];
@@ -202,39 +198,6 @@ export class HttpClient implements Server.IHttpClient {
 
 			result.resolve(finalResult);
 		}
-	}
-
-	private trackDownloadProgress(pipeTo: NodeJS.WritableStream): NodeJS.ReadableStream {
-		// \r for carriage return doesn't work on windows in node for some reason so we have to use it's hex representation \x1B[0G
-		let lastMessageSize = 0;
-		const carriageReturn = "\x1B[0G";
-		let timeElapsed = 0;
-
-		const progressStream = progress({ time: 1000 }, (progress: any) => {
-			timeElapsed = progress.runtime;
-
-			if (timeElapsed >= 1) {
-				this.$logger.write("%s%s", carriageReturn, Array(lastMessageSize + 1).join(" "));
-
-				const message = util.format("%sDownload progress ... %s | %s | %s/s",
-					carriageReturn,
-					Math.floor(progress.percentage) + "%",
-					filesize(progress.transferred),
-					filesize(progress.speed));
-
-				this.$logger.write(message);
-				lastMessageSize = message.length;
-			}
-		});
-
-		progressStream.on("finish", () => {
-			if (timeElapsed >= 1) {
-				this.$logger.out("%s%s%s%s", carriageReturn, Array(lastMessageSize + 1).join(" "), carriageReturn, "Download completed.");
-			}
-		});
-
-		progressStream.pipe(pipeTo);
-		return progressStream;
 	}
 
 	private getErrorMessage(statusCode: number, body: string): string {
